Estimate ecological quantities

Produces estimates of overall conditional means from a fitted ecological inference model or Riesz representer. If both a regression model and a Riesz representer are provided, a debiased machine learning (DML) estimate is produced.

Usage

ei_est(
  regr = NULL,
  riesz = NULL,
  data,
  total,
  outcome = NULL,
  conf_level = FALSE
)

# S3 method for class 'ei_est'
as.matrix(x, se = FALSE, ...)

# S3 method for class 'ei_est'
vcov(object, ...)

Arguments

regr

A fitted regression model, from ei_ridge(), or another kind of regression model wrapped with ei_wrap_model(). If riesz is not provided and regr is an ei_riesz() object, then riesz will be set to the value of regr and regr will be set to NULL. This is so users can call this function as ei_est(<riesz>, data = <data>).

riesz

A fitted Riesz representer, from ei_riesz(), or a matrix of Riesz weights

data

The data frame, matrix, or ei_spec object that was used to fit the regression or Riesz representer.

total

<tidy-select> A variable containing the total number of observations in each aggregate unit. For example, the column containing the total number of voters. Required if data is not an ei_spec() object and riesz is not provided.

outcome

<data-masking> A vector or matrix of outcome variables. Only required if both riesz is provided alone (without regr) and data is not an ei_spec object.

conf_level

A numeric specifying the level for confidence intervals. If FALSE (the default), no confidence intervals are calculated. Standard errors are always returned.

x, object

An object of class ei_est

se

If TRUE, return standard errors instead of estimates.

...

Additional arguments (ignored)

Value

A data frame with estimates. It has two additional attributes: vcov, containing the estimated covariance matrix for the estimates, and n, containing the number of aggregate units used in estimation (the number of rows in data).

Methods (by generic)

as.matrix(ei_est): Format estimates or standard errors as a matrix

vcov(ei_est): Extract full covariance matrix of estimates

Examples

data(elec_1968)

spec = ei_spec(elec_1968, vap_white:vap_other, pres_dem_hum:pres_oth,
               total = pres_total, covariates = c(pop_urban, farm))

m = ei_ridge(spec)
rr = ei_riesz(spec, penalty = m$penalty)

ei_est(regr = m, data = spec, conf_level = 0.95) # Plug-in estimate
#> # A tibble: 15 × 6
#>    predictor outcome estimate std.error conf.low conf.high
#>    <chr>     <chr>      <dbl>     <dbl>    <dbl>     <dbl>
#>  1 white     dem_hum  0.248    0.0154    0.218     0.278  
#>  2 black     dem_hum  0.420    0.0381    0.345     0.495  
#>  3 other     dem_hum  3.87     0.477     2.93      4.80   
#>  4 white     rep_nix  0.387    0.0271    0.334     0.440  
#>  5 black     rep_nix -0.0894   0.00506  -0.0993   -0.0795 
#>  6 other     rep_nix  1.40     0.233     0.943     1.85   
#>  7 white     ind_wal  0.363    0.0279    0.309     0.418  
#>  8 black     ind_wal  0.667    0.0418    0.585     0.749  
#>  9 other     ind_wal -4.25     0.595    -5.41     -3.08   
#> 10 white     abs      0.00153  0.000179  0.00118   0.00188
#> 11 black     abs      0.00278  0.000290  0.00221   0.00334
#> 12 other     abs     -0.0188   0.00233  -0.0233   -0.0142 
#> 13 white     oth      0.00153  0.000179  0.00118   0.00188
#> 14 black     oth      0.00278  0.000290  0.00221   0.00334
#> 15 other     oth     -0.0188   0.00233  -0.0233   -0.0142 
ei_est(riesz = rr, data = spec) # Weighted (Riesz) estimate
#> # A tibble: 15 × 4
#>    predictor outcome estimate std.error
#>    <chr>     <chr>      <dbl>     <dbl>
#>  1 white     dem_hum  0.248    0.0213  
#>  2 black     dem_hum  0.420    0.0796  
#>  3 other     dem_hum  3.87     2.36    
#>  4 white     rep_nix  0.387    0.0211  
#>  5 black     rep_nix -0.0894   0.0587  
#>  6 other     rep_nix  1.40     1.72    
#>  7 white     ind_wal  0.363    0.0214  
#>  8 black     ind_wal  0.667    0.0830  
#>  9 other     ind_wal -4.25     1.78    
#> 10 white     abs      0.00153  0.000282
#> 11 black     abs      0.00278  0.000858
#> 12 other     abs     -0.0188   0.0173  
#> 13 white     oth      0.00153  0.000282
#> 14 black     oth      0.00278  0.000858
#> 15 other     oth     -0.0188   0.0173  
est = ei_est(regr = m, riesz = rr, data = spec) # Double/debiased ML estimate
as.matrix(est)
#>         dem_hum     rep_nix    ind_wal          abs          oth
#> white 0.2481598  0.38699489  0.3633165  0.001528811  0.001528811
#> black 0.4200405 -0.08937756  0.6665617  0.002775393  0.002775393
#> other 3.8671722  1.39827216 -4.2466833 -0.018760984 -0.018760984
as.matrix(est, se=TRUE)
#>          dem_hum    rep_nix    ind_wal          abs          oth
#> white 0.01770099 0.02909194 0.02826466 0.0002937972 0.0002937972
#> black 0.04602302 0.02328971 0.05257540 0.0008645117 0.0008645117
#> other 0.78202192 0.53981237 1.09745695 0.0158211423 0.0158211423
vcov(est)[1:4, 1:4]
#>               white:dem_hum black:dem_hum other:dem_hum white:rep_nix
#> white:dem_hum  0.0003133249  0.0002732986   0.006420237  0.0004300809
#> black:dem_hum  0.0002732986  0.0021181180   0.011485449  0.0007930489
#> other:dem_hum  0.0064202365  0.0114854489   0.611558291  0.0129493462
#> white:rep_nix  0.0004300809  0.0007930489   0.012949346  0.0008463410