Specify an ecological inference problem

Uses tidy-select syntax to specify outcomes, predictors, and covariates. The result of this function can be passed directly into ei_ridge() or ei_riesz(), or plotted with plot().

Usage

ei_spec(data, predictors, outcome, weights, covariates = NULL, strip = TRUE)

# S3 method for class 'ei_spec'
weights(object, normalize = TRUE, ...)

Arguments

data

A data frame.

predictors

<tidy-select> Predictor variables. This is the x variable in ecological regression that is of primary interest. For example, the columns containing the percentage of each racial group.

outcome

<tidy-select> Outcome variables. This is the y variable in ecological regression that is of primary interest. For example, the columns containing the percentage of votes for each party.

weights

<data-masking> A vector of unit weights. In general these should be the count of the total number of individuals in each unit. Required by default. To force constant weights, you can provide weights=FALSE.

covariates

<tidy-select> Covariates.

strip

Whether to strip common prefixes from column names within each group. For example, columns named vap_white, vap_black, and vap_hisp would be renamed white, black and other in the model and output.

object

An ei_spec object.

normalize

If TRUE, normalize the weights to have mean 1.

...

Additional arguments (ignored).

Value

An ei_spec object, which is a data frame with additional attributes recording predictors, outcomes, and covariates.

Details

The function is lightweight and does not perform any checking of the arguments, bounds, sum constraints, etc. All of these checks are performed by ei_ridge() or ei_riesz().

Methods (by generic)

weights(ei_spec): Extract the weights from a specification

Examples

data(elec_1968)
ei_spec(elec_1968, vap_white:vap_other, pres_dem_hum:pres_oth, pres_total)
#> EI Specification
#> • Predictors: `white`, `black`, and `other`
#> • Outcome: `dem_hum`, `rep_nix`, `ind_wal`, `abs`, and `oth`
#> • Covariates: none
#> # A tibble: 1,143 × 8
#>   white  black    other dem_hum rep_nix ind_wal     abs     oth
#>   <dbl>  <dbl>    <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1 0.761 0.237  0.00173   0.199   0.0773   0.711 0.0122  0.0122 
#> 2 0.860 0.137  0.00306   0.105   0.115    0.764 0.0161  0.0161 
#> 3 0.610 0.389  0.000808  0.242   0.0489   0.687 0.0218  0.0218 
#> 4 0.783 0.216  0.00106   0.141   0.0571   0.799 0.00290 0.00290
#> 5 0.981 0.0181 0.000757  0.0375  0.222    0.727 0.0134  0.0134 
#> # ℹ 1,138 more rows