Estimate ecological quantities

Produces estimates of overall conditional means from a fitted ecological inference model or Riesz representer. If both a regression model and a Riesz representer are provided, a debiased machine learning (DML) estimate is produced.

Usage

ei_est(
  regr = NULL,
  riesz = NULL,
  data,
  weights,
  outcome = NULL,
  conf_level = FALSE
)

# S3 method for class 'ei_est'
as.matrix(x, se = FALSE, ...)

# S3 method for class 'ei_est'
vcov(object, ...)

Arguments

regr

A fitted regression model, from ei_ridge(). If riesz is not provided and regr is an ei_riesz() object, then riesz will be set to the value of regr and regr will be set to NULL. This is so users can call this function as ei_est(<riesz>, data = <data>).

riesz

A fitted Riesz representer, from ei_riesz(), or a matrix of Riesz weights

data

The data frame, matrix, or ei_spec object that was used to fit the regression or Riesz representer.

weights

<data-masking> A vector of unit weights. In general these should be the count of the total number of individuals in each unit. Required by default. To force constant weights, you can provide weights=FALSE.

outcome

<data-masking> A vector or matrix of outcome variables. Only required if both riesz is provided alone (without regr) and data is not an ei_spec object.

conf_level

A numeric specifying the level for confidence intervals. If FALSE (the default), no confidence intervals are calculated. Standard errors are always returned.

x, object

An object of class ei_est

se

If TRUE, return standard errors instead of estimates.

...

Additional arguments (ignored)

Value

A data frame with estimates.

Methods (by generic)

as.matrix(ei_est): Format estimates or standard errors as a matrix

vcov(ei_est): Extract full covariance matrix of estimates

Examples

data(elec_1968)

spec = ei_spec(elec_1968, vap_white:vap_other, pres_dem_hum:pres_oth,
               weights = pres_total, covariates = c(pop_urban, farm))

m = ei_ridge(spec)
rr = ei_riesz(spec, penalty = m$penalty)

ei_est(regr = m, data = spec, conf_level = 0.95) # Plug-in estimate
#> # A tibble: 15 × 6
#>    predictor outcome estimate std.error   conf.low conf.high
#>    <chr>     <chr>      <dbl>     <dbl>      <dbl>     <dbl>
#>  1 white     dem_hum  0.251    0.0207     0.211      0.292  
#>  2 black     dem_hum  0.382    0.0318     0.320      0.445  
#>  3 other     dem_hum  8.20     1.06       6.12      10.3    
#>  4 white     rep_nix  0.448    0.0350     0.380      0.517  
#>  5 black     rep_nix -0.154    0.00916   -0.172     -0.136  
#>  6 other     rep_nix  1.01     0.188      0.638      1.38   
#>  7 white     ind_wal  0.299    0.0166     0.267      0.332  
#>  8 black     ind_wal  0.767    0.0518     0.666      0.869  
#>  9 other     ind_wal -8.15     1.14     -10.4       -5.93   
#> 10 white     abs      0.00115  0.000234   0.000693   0.00161
#> 11 black     abs      0.00436  0.000515   0.00335    0.00537
#> 12 other     abs     -0.0522   0.00663   -0.0652    -0.0392 
#> 13 white     oth      0.00115  0.000234   0.000693   0.00161
#> 14 black     oth      0.00436  0.000515   0.00335    0.00537
#> 15 other     oth     -0.0522   0.00663   -0.0652    -0.0392 
ei_est(riesz = rr, data = spec) # Weighted (Riesz) estimate
#> # A tibble: 15 × 4
#>    predictor outcome estimate std.error
#>    <chr>     <chr>      <dbl>     <dbl>
#>  1 white     dem_hum  0.251    0.0471  
#>  2 black     dem_hum  0.382    0.186   
#>  3 other     dem_hum  8.20     5.07    
#>  4 white     rep_nix  0.448    0.0574  
#>  5 black     rep_nix -0.154    0.174   
#>  6 other     rep_nix  1.01     5.20    
#>  7 white     ind_wal  0.299    0.0329  
#>  8 black     ind_wal  0.767    0.154   
#>  9 other     ind_wal -8.15     3.26    
#> 10 white     abs      0.00115  0.000377
#> 11 black     abs      0.00436  0.00208 
#> 12 other     abs     -0.0522   0.0558  
#> 13 white     oth      0.00115  0.000377
#> 14 black     oth      0.00436  0.00208 
#> 15 other     oth     -0.0522   0.0558  
est = ei_est(regr = m, riesz = rr, data = spec) # Double/debiased ML estimate
as.matrix(est)
#>         dem_hum    rep_nix    ind_wal          abs          oth
#> white 0.2514755  0.4480782  0.2992944  0.001151928  0.001151928
#> black 0.3824496 -0.1538928  0.7670800  0.004363138  0.004363138
#> other 8.2002756  1.0065979 -8.1546861 -0.052187355 -0.052187355
as.matrix(est, se=TRUE)
#>          dem_hum    rep_nix    ind_wal          abs          oth
#> white 0.02392274 0.03679981 0.02037134 0.0003795963 0.0003795963
#> black 0.05186651 0.03741616 0.07577585 0.0020152257 0.0020152257
#> other 1.81579169 0.98068420 2.19031071 0.0505327951 0.0505327951
vcov(est)[1:4, 1:4]
#>               white:dem_hum black:dem_hum other:dem_hum white:rep_nix
#> white:dem_hum  0.0005722975 -0.0000398951   0.020253249  0.0007052276
#> black:dem_hum -0.0000398951  0.0026901346   0.002987941  0.0005413884
#> other:dem_hum  0.0202532487  0.0029879409   3.297099467  0.0382632050
#> white:rep_nix  0.0007052276  0.0005413884   0.038263205  0.0013542260