Generate synthetic ecological data

Samples data from a following truncated Normal ecological model. The data can be generated completely at random, or can be generated conditional on provided predictors x and/or covariates z.

Usage

ei_synthetic(
  n,
  p = 0,
  n_x = 2,
  x = n_x:1,
  z = 0.25 * exp(-(seq_len(p) - 1)/2),
  r2_xz = 0.5,
  r2_bz = 0.5,
  b_loc = NULL,
  b_cov = NULL
)

Arguments

n

The number of rows (geographies) to generate. Defaults to the number of rows in x or z, if they are a matrix or data frame.

p

The number of covariates. Defaults to the number of columns in z, if it is a matrix or data frame, or the length of z, if it is a vector of singular values.

n_x

The number of predictor variables. Defaults to the number of columns in x, if it is a matrix or data frame, or the length of x, if it is a vector of mean parameters for the softmax-transformed Normal distribution.

x

Either a matrix or data frame containing the predictor percentages in each row, or a vector containing Dirichlet parameters to use in sampling predictor percentages.

z

A matrix or data frame containing geography-level covariates, or a vector of values to form a Toeplitz covariance matrix for the random covariates.

r2_xz

The approximate \(R^2\) of the covariates z and predictors x. See the model specification for details. If either r2_xz or r2_bz are zero, then there is no confounding, and an unadjusted Goodman regression will estimate the global parameters correctly.

r2_bz

The approximate \(R^2\) of the covariates z and unit-level parameters b. See the model specification for details. If either r2_xz or r2_bz are zero, then there is no confounding, and an unadjusted Goodman regression will estimate the global parameters correctly.

b_loc

The center of the distribution of geography-level parameters. Defaults to a linearly spaced sequence across groups from 0.5 to 0.9. Because of the truncation, this will not exactly be the mean of the geography-level parameters.

b_cov

The residual covariance matrix for geography-level parameters. Defaults to 0.02 * (1 + diag(n_x)).

Value

An ei_spec object with additional attributes:

b_loc and b_cov

Lambda with the coefficients of z

eta, the linear predictor for b

est_true, the mean of the geography-level parameters, formatted similarly to the output from ei_est()

r2_xz_act and r2_bz_act, containing the actual (sample) \(R^2\) values for x and z, and b and z, respectively.

Details

This function samples data from the following truncated Normal ecological model: $$ \begin{pmatrix}x_i\\ z_i\end{pmatrix} \stackrel{\text{iid}}{\sim} \mathcal{N}_{[0,1]^{n_x} \times \mathbb{R}^p}\left( \begin{pmatrix}\mu_x\\ 0\end{pmatrix}, \begin{pmatrix}\Sigma_x & \Gamma \\ \Gamma & T\end{pmatrix}\right) $$ $$ \eta = z_i^\top \Lambda + \mathtt{b_{loc}} $$ $$ b_i \stackrel{\text{iid}}{\sim} \mathcal{N}_{[0, 1]^{n_x}}(\eta, \mathtt{B_{cov}}) $$ $$ y_i = b_i^\top x_i, $$ where \(\mu_x\) and \(\Sigma_x\) are the mean and covariance of the Normal approximation to a Dirichlet distribution with parameters supplied by the x argument below, and \(\Gamma\), \(T\), and \(\Gamma\) are matrices sampled to have certain properties, as described below. The subscripts on \(\mathcal{N}\) indicate truncation; i.e., both the predictors x and the unit-level parameters b are truncated to the n_x-dimensional hypercube.

The matrix \(T\) is a symmetric Toeplitz matrix with diagonals provided by the z argument. Generally, a decreasing set of nonnegative values will be sufficient for a positive definite \(T\).

The matrices \(\Gamma\) and \(\Lambda\) are initially filled with independent samples from a standard Normal distribution. \(Gamma\) is then projected so that its rows sum to zero, preserving the sum-to-1 requirement on x, and so that its columns are scaled to produce the correct \(R^2\) value matching r2_xz. The matrix \(\Lambda\) is likewise scaled to produce the correct \(R^2\) value matching r2_bz. Due to the truncation in the sampling of x and b, the in-sample \(R^2\) values will generally be slightly smaller than the provided arguments.

Aspects of the model can be replaced with data provided to the function. If x or z is provided as a matrix or data frame, then the other value is sampled from its marginal distribution. If both are provided, then the first row of the model is skipped.

Examples

ei_synthetic(n = 10)
#> EI Specification
#> • Predictors: `x1` and `x2`
#> • Outcome: `y`
#> • Covariates: none
#> # A tibble: 10 × 3
#>       y    x1      x2
#>   <dbl> <dbl>   <dbl>
#> 1 0.288 0.940 0.0600 
#> 2 0.884 0.535 0.465  
#> 3 0.517 0.433 0.567  
#> 4 0.579 0.996 0.00377
#> 5 0.263 0.807 0.193  
#> # ℹ 5 more rows

ei_synthetic(n = 10, p = 2, n_x = 3)
#> EI Specification
#> • Predictors: `x1`, `x2`, and `x3`
#> • Outcome: `y`
#> • Covariates: `z1` and `z2`
#> # A tibble: 10 × 6
#>       y    x1     x2    x3     z1     z2
#>   <dbl> <dbl>  <dbl> <dbl>  <dbl>  <dbl>
#> 1 0.432 0.209 0.380  0.411  1.22   0.995
#> 2 0.806 0.583 0.0825 0.334  0.196 -0.489
#> 3 0.744 0.580 0.259  0.161 -0.410 -0.836
#> 4 0.556 0.576 0.203  0.220  0.302  0.163
#> 5 0.543 0.526 0.205  0.268  0.687  0.364
#> # ℹ 5 more rows

# Manual hyperparameters: x2 dominant and z1, z2 very correlated
ei_synthetic(n = 10, x = c(1, 95, 4), z = c(10, 9.999))
#> EI Specification
#> • Predictors: `x1`, `x2`, and `x3`
#> • Outcome: `y`
#> • Covariates: `z1` and `z2`
#> # A tibble: 10 × 6
#>        y       x1    x2     x3     z1     z2
#>    <dbl>    <dbl> <dbl>  <dbl>  <dbl>  <dbl>
#> 1 0.994  0.00320  0.961 0.0355 -0.369 -0.340
#> 2 0.931  0.000921 0.946 0.0529  0.787  0.878
#> 3 0.933  0.00633  0.965 0.0284 -2.63  -2.65 
#> 4 0.0342 0.00763  0.960 0.0323  2.68   2.66 
#> 5 0.639  0.00883  0.974 0.0174 -1.20  -1.20 
#> # ℹ 5 more rows

# Condition on provided x but not z
data(elec_1968)
ei_synthetic(
    x = cbind(elec_1968$pop_white, 1 - elec_1968$pop_white),
    p = 5,
    b_loc = c(0.3, 0.9),
    b_cov = matrix(c(0.02, 0.016, 0.016, 0.2), nrow=2)
)
#> EI Specification
#> • Predictors: `x1` and `x2`
#> • Outcome: `y`
#> • Covariates: `z1`, `z2`, `z3`, `z4`, and `z5`
#> # A tibble: 1,143 × 8
#>       y    x1     x2     z1     z2     z3      z4     z5
#>   <dbl> <dbl>  <dbl>  <dbl>  <dbl>  <dbl>   <dbl>  <dbl>
#> 1 0.329 0.716 0.284   0.531  0.268  0.580 -0.0719 -0.508
#> 2 0.312 0.819 0.181  -0.709 -0.441 -0.405 -0.182  -0.456
#> 3 0.625 0.538 0.462   0.330  0.725  0.417  0.603   0.640
#> 4 0.252 0.721 0.279  -0.461 -1.30  -0.741 -0.0416  0.108
#> 5 0.159 0.976 0.0241  0.176  0.458  1.02   0.156   0.458
#> # ℹ 1,138 more rows