Generate synthetic ecological data

Samples data from the following truncated Normal ecological model: $$ x_i \stackrel{\text{iid}}{\sim} \mathrm{Dirichlet}(\alpha) \\ \eta = z_i^\top \mathtt{coefs} + \mathtt{b_{loc}} \\ b_i \stackrel{\text{iid}}{\sim} \mathcal{N}_{[0, 1]^d}(\eta, \mathtt{B_{cov}}) \\ y_i = b_i^\top x_i, $$ where \(\alpha\) is a vector of hyperparameters that may be optionally specified as the x argument to the function, and \(\mathcal{N}_{[0, 1]^d}\) indicates a multivariate Normal distribution truncated to the d-dimensional hypercube. Aspects of the model can be replaced with data provided to the function.

Usage

ei_synthetic(
  n,
  p = 0,
  n_x = 2,
  x = n_x:1,
  z = NULL,
  b_loc = NULL,
  coefs = NULL,
  b_cov = NULL
)

Arguments

n

The number of rows (geographies) to generate. Defaults to the number of rows in x, if it is a matrix or data frame.

p

The number of covariates. Defaults to the number of columns in z, if it is a matrix or data frame, or the length of z, if it is a vector of singular values.

n_x

The number of predictor variables. Defaults to the number of columns in x, if it is a matrix or data frame, or the length of x, if it is a vector of Dirichlet parameters.

x

Either a matrix or data frame containing the predictor percentages in each row, or a vector containing Dirichlet parameters to use in sampling predictor percentages.

z

A matrix or data frame containing geography-level covariates, or a vector of singular values to use in generating the random covariates. Singular values default to exp(-(seq_len(p) - 1)/2). They are applied to scale a random orthogonal matrix.

b_loc

The center of the distribution of geography-level parameters. Defaults to a linearly spaced sequence across groups from 0.5 to 0.9. Because of the truncation, this will not exactly be the mean of the geography-level parameters.

coefs

Either A p-by-n_x matrix of coefficients for z, or The product with z is added to ctr to form the mean of the individual geography-level parameters. Defaults to a matrix of standard Normals.

b_cov

The covariance matrix for geography-level parameters. Defaults to 0.02 * (1 + diag(n_x)).

Value

An ei_spec object with additional attributes b_loc, coefs, and b_cov, eta (the linear predictor for b), and est_true (the mean of the geography-level parameters, formatted similarly to the output from ei_est()).

Examples

ei_synthetic(n = 10)
#> EI Specification
#> • Predictors: `x1` and `x2`
#> • Outcome: `y`
#> • Covariates: none
#> # A tibble: 10 × 3
#>       y    x1    x2
#>   <dbl> <dbl> <dbl>
#> 1 0.826 0.233 0.767
#> 2 0.438 0.827 0.173
#> 3 0.708 0.763 0.237
#> 4 0.690 0.524 0.476
#> 5 0.460 0.852 0.148
#> # ℹ 5 more rows

ei_synthetic(n = 10, p = 2, n_x = 3)
#> EI Specification
#> • Predictors: `x1`, `x2`, and `x3`
#> • Outcome: `y`
#> • Covariates: `z1` and `z2`
#> # A tibble: 10 × 6
#>       y    x1    x2     x3     z1      z2
#>   <dbl> <dbl> <dbl>  <dbl>  <dbl>   <dbl>
#> 1 0.921 0.722 0.171 0.107  -0.170  0.106 
#> 2 0.427 0.315 0.601 0.0843  0.329 -0.220 
#> 3 0.197 0.628 0.154 0.217   0.672 -0.142 
#> 4 1.18  0.253 0.591 0.156  -0.232 -0.0341
#> 5 1.08  0.677 0.184 0.139  -0.438 -0.416 
#> # ℹ 5 more rows

data(elec_1968)
ei_synthetic(
    x = cbind(elec_1968$pop_white, 1 - elec_1968$pop_white),
    b_loc = c(0.3, 0.9),
    b_cov = matrix(c(0.02, 0.016, 0.016, 0.2), nrow=2)
)
#> EI Specification
#> • Predictors: `x1` and `x2`
#> • Outcome: `y`
#> • Covariates: none
#> # A tibble: 1,143 × 3
#>       y    x1     x2
#>   <dbl> <dbl>  <dbl>
#> 1 0.378 0.716 0.284 
#> 2 0.394 0.819 0.181 
#> 3 0.437 0.538 0.462 
#> 4 0.168 0.721 0.279 
#> 5 0.188 0.976 0.0241
#> # ℹ 1,138 more rows