Estimation weighting models

Several built-in helper functions to generate estimation weights from a vector of unit totals, or an existing ei_spec() object.

Usage

ei_wgt_unif(x)

ei_wgt_prop(x)

ei_wgt_sqrt(x)

Arguments

x

A numeric vector of unit totals, or an existing ei_spec() object.

Value

A numeric vector of estimation weights with the same number of observations as x. These will have mean 1.

Functions

ei_wgt_unif(): Uniform weights. Appropriate if the unit-level variance is constant, i.e., homosekdastic.

ei_wgt_prop(): Weights proportional to the totals. Appropriate if the unit-level variance is inversely proportional to the number of observations.

ei_wgt_sqrt(): Weights proportional to the square root of the totals. Appropriate if the unit-level variance is inversely proportional to the square root of the number of observations.

Examples

data(elec_1968)

ei_wgt_unif(elec_1968$pres_total)
#>    [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>   [38] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>   [75] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [112] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [149] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [186] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [223] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [260] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [297] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [334] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [371] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [408] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [445] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [482] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [519] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [556] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [593] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [630] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [667] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [704] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [741] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [778] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [815] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [852] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [889] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [926] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [963] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [1000] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [1037] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [1074] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [1111] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1

spec = ei_spec(elec_1968, vap_white:vap_other, pres_ind_wal, total=pres_total)
ei_wgt_prop(spec)
#>    [1]  0.596840482  1.450706556  0.617016024  0.351951135  0.693389343
#>    [6]  0.335563338  0.565688207  2.077153294  0.822327295  0.437600566
#>   [11]  0.683881329  0.473313595  0.670972074  0.386736554  0.307039295
#>   [16]  0.838637791  1.198705524  0.405134175  0.287404858  1.011559972
#>   [21]  0.423145292  1.331122017  0.746649684  1.282190528  1.139956817
#>   [26]  0.913465092  0.833613231  2.388134936  0.475787225  0.708617627
#>   [31]  0.663937689  0.306807392  0.399723110  0.407066698  1.325324447
#>   [36]  0.811273262 15.452533239  0.456771196  1.458436649  0.592588931
#>   [41]  1.014188204  0.805011887  0.251459926  0.493952944  4.008903539
#>   [46]  0.701660543  0.642834535  1.283736547  7.133097932  0.566538517
#>   [51]  3.588463780  1.699770153  0.426778435  0.499595912  0.680170884
#>   [56]  0.527037742  0.870408474  0.750205527  0.830984999  0.371430969
#>   [61]  1.437797301  0.905812300  2.203617616  1.507368138  0.437832468
#>   [66]  0.345921662  0.427860648  0.587487069  0.611141154  0.530748186
#>   [71]  1.254439495  0.573959406  0.371430969  0.169289037  0.391761114
#>   [76]  0.436518353  0.552778952  0.491479314  0.321571869  0.190933297
#>   [81]  0.637423470  0.501760338  1.200251543  0.557958114  0.818307647
#>   [86]  0.440924506  0.281916492  0.441852117  0.361072645  0.844280759
#>   [91]  0.355043172  0.254861167  1.567972067  0.283926316  0.624359613
#>   [96]  0.559736035  0.622736293  0.309744827  0.606116593  0.230975179
#>  [101]  0.535850048  1.055853405  0.394775850  0.277046534  0.480347980
#>  [106]  0.376919335  0.292274817  0.256097982  0.502378745  0.594676056
#>  [111]  0.380475178  0.251382625  0.823486809  1.198860126  0.385963544
#>  [116]  0.195262150  0.303096947  0.223090484  0.915474916  0.177019130
#>  [121]  0.886796271  0.254706565  0.609595135  0.401655633  0.669889861
#>  [126]  0.276891932  6.195669552  0.325746120  0.706994307  0.873036705
#>  [131]  0.262823163  0.259653824  2.090371753  0.297376678  0.267461218
#>  [136]  0.205929678  1.329421397  0.285858840  1.690030236  0.938355991
#>  [141]  0.280525076  0.408226212  2.197356241  0.212036451  1.878567205
#>  [146]  0.365710701  5.976444114 15.052578226  0.241874610  0.925524037
#>  [151]  0.552547049  0.715110905  0.815061008  0.568084536 28.238184387
#>  [156]  0.316470008  0.161404342 12.907554715  5.285064595  0.137441054
#>  [161]  0.242493018  0.700114524  0.116724405  0.084335315  0.293743535
#>  [166]  0.211108840  0.348627195  0.257334796  0.461022747  0.820703975
#>  [171] 10.992346870  0.416574713  0.983190531  0.953120469  0.239014476
#>  [176]  0.114173474  1.917372272  2.403672423  2.520474129  0.306188984
#>  [181]  0.132802998  0.366019904  2.685975420  1.767563069  0.790788515
#>  [186]  1.151706558  0.505006977  1.609482667  0.232521198  7.781807338
#>  [191]  0.734668040  9.038179355  1.777998694 16.330439902  5.818982119
#>  [196]  0.852242755  0.874273520  1.308163641  0.982958628  3.658498422
#>  [201]  1.871919325  0.391606512  0.462414164  0.390601600  0.128396845
#>  [206]  5.428071315  0.182043691  0.553629262  0.381248188  0.327214837
#>  [211]  0.195416751  0.216442604  0.132493794  0.549686914  0.164496379
#>  [216]  0.399877712  0.637809975  0.260581436  0.295907961  3.200026606
#>  [221]  0.279133659  0.174932005  0.292197516  0.183125904  0.607121505
#>  [226]  0.378310752  0.234453721  0.147644777  0.300314114  0.213582470
#>  [231]  0.925292134  0.726087637  0.150273008  4.139078305  0.040273785
#>  [236]  0.414796791  0.576819541  1.144672174  0.097244570  1.811779201
#>  [241]  0.137595656  3.494697751  0.491402013  0.720908475  0.367024816
#>  [246]  0.275500515  0.574887018  0.125304808  0.403742758  0.259344621
#>  [251]  0.123681488  0.545280761  8.057617056  0.439919594  0.242415717
#>  [256]  1.450320052  0.560354443  0.302169336  0.049627197  0.306498188
#>  [261]  0.416033606  0.472463285  0.194025335  0.455457080  0.255634176
#>  [266]  1.662897609  0.342675023  0.322576782 13.842122961  0.310981642
#>  [271]  0.074595398  0.952115557  0.467902530  0.448577298  0.271326265
#>  [276]  1.351915967  0.439764992  1.054616590  0.282148395  0.423067991
#>  [281]  0.304874869  0.369034641  0.140069285  0.613614784  1.117539547
#>  [286]  0.221080660  0.475323420  0.171376162  0.225023008  0.403278953
#>  [291]  0.195339450  0.221699068  0.275809719  0.216829109  0.119198034
#>  [296]  0.992389341  0.175009306  0.272794983  0.169211736  0.136590744
#>  [301]  0.862214575  0.164728282  0.311136244  0.162254652  0.240483194
#>  [306]  0.289955789  0.099099792  0.421367370  0.176400723  0.435513440
#>  [311]  0.248908995  0.176864528  0.230356772  0.297299377  2.673839174
#>  [316]  0.515983709  0.195725955  0.201214321  0.390678901  0.301782831
#>  [321]  0.288177868  0.249682004  0.186990950  0.616552219  0.207011891
#>  [326]  0.212036451  0.057743795  0.207089192  0.229429161  2.806178366
#>  [331]  0.357362200  0.084412616  0.321339967  0.192711219  0.848686912
#>  [336]  0.396708374  0.127855738  0.508099014  0.117110909  0.109612719
#>  [341]  0.403047050  0.209485521  0.329301962  0.279752066  0.764970005
#>  [346]  0.527269645  0.440460700  0.220384952  0.146562564  0.955980603
#>  [351]  0.206857289  0.178951653  0.239710184  0.508099014  1.107026621
#>  [356]  0.540951909  0.788237585  0.135663132  0.364783089  0.441774816
#>  [361]  0.055115563  0.129324456  0.182043691  0.889424502  0.206238882
#>  [366]  0.273258788  0.261586348  0.337959667  1.313420104  0.561127452
#>  [371]  0.815138308  0.479652272  0.942452940  0.559040327  0.477333244
#>  [376]  1.219499474  5.207145257  3.429301164  0.287172956  0.200441312
#>  [381]  0.324741208  0.461718456  0.579679675  0.662005166  6.082114485
#>  [386]  0.326055323  0.316238105  0.893444151  0.550923729  0.426314630
#>  [391]  1.470959400  0.756544203  0.507867111  7.078678077  0.753761370
#>  [396]  2.348015753  1.640557641  0.451901238  0.686200357  0.947245598
#>  [401]  0.439687691  0.691224917  0.912305578 13.811202589  2.450748690
#>  [406]  0.660304545  0.579525073  2.723389070  0.291501808  0.499595912
#>  [411]  0.526419334  1.470804798  0.705602890  0.260504135  0.504775074
#>  [416]  0.574423212  1.951230079  0.750205527  1.287756195  1.604844611
#>  [421]  1.621618913  0.203919854  1.443749472  0.521472075  1.175592546
#>  [426]  0.665251805  1.215016020  1.083217934  0.406139087  0.352028436
#>  [431]  0.206548085  0.486609355  1.043639858  0.710086344  0.396708374
#>  [436]  0.538246377  0.206006979  0.889269900  0.424614009  0.246899171
#>  [441]  0.399104702  0.245121250  0.270707857  0.416652014  0.425850824
#>  [446]  0.842425537  0.647704494  0.371353668  0.644380554  1.254284893
#>  [451]  0.266069802  0.338346171  0.257025593  0.549068507  0.466974919
#>  [456]  2.260897605  4.694794692  0.572722592  0.280447775  0.085417528
#>  [461]  0.478492758  1.579953711  0.344762148  0.261277144  0.338268870
#>  [466]  1.390952937  0.259112718  0.474782313  0.411163647  1.574233443
#>  [471]  0.301009822  0.488232675  1.056394511  0.899164420  0.766670625
#>  [476]  0.852320056  0.731421401  0.644148651  0.485295240  0.813901494
#>  [481]  0.336954755  0.604106769  0.533531020  0.282844103  0.558808424
#>  [486]  0.616397617  0.639587896  0.247904083  0.784913645  0.473854702
#>  [491]  0.480657184  0.338500773  0.952579362  0.522863492  0.186217941
#>  [496]  0.542497928  0.398563596  0.209640123  0.585168041  0.396553772
#>  [501]  0.353806357  0.454452168  0.428556357  0.180033866  0.523327297
#>  [506]  0.371508270  1.090870726  1.184250250  0.392302221  0.310981642
#>  [511]  0.302942345  0.467979831  0.321571869  0.623045497  2.604268337
#>  [516]  0.650564628  0.286090743  0.619489654  0.711786965  0.349245602
#>  [521]  0.895685878  0.550846428  0.649095911  0.675146324  3.675195423
#>  [526]  1.751948281  1.953008000  1.567199058  0.153596948  0.882467419
#>  [531]  0.465660803  2.524029971  0.820626675  0.547677090  0.285626937
#>  [536]  0.195571353  1.747619429  1.145367882  1.062114780  2.212352621
#>  [541]  0.198817992  0.199359099  2.768610114  0.609363232  0.947477501
#>  [546]  3.309252820  1.105557903  5.224383364  0.754070574  3.309562024
#>  [551]  0.215205790  0.231438985  0.660613749  0.395471559  6.520410759
#>  [556]  1.174278430  1.215325224  1.221818502  1.255985513  0.510418042
#>  [561]  0.351100825  0.154833763  1.890471548  0.601633139  1.582195438
#>  [566]  0.260194931  0.681871505  1.216175534  1.035291358  0.796276882
#>  [571]  0.504543171  0.492020420  0.630543687  8.309618089  0.401964837
#>  [576]  0.598231898  0.940597718  1.477607280  2.091840471  0.612068765
#>  [581]  0.948250510  1.415225429  0.268388829  0.586791361  0.438218973
#>  [586]  0.235381332  0.683881329  1.747619429  0.429561269  1.986092798
#>  [591]  0.972368400  1.485337373  1.870141403  2.512357531  1.305071604
#>  [596]  1.230708109  0.462646067  1.417157953  0.816684327  1.455499214
#>  [601]  0.251846430  0.665406407  0.099486297  1.057554025  0.877210955
#>  [606]  5.191298566  0.416110907  0.369498446  0.702897358  1.524760847
#>  [611]  1.435323671  1.246786702  0.751751546  0.418584537  0.451360131
#>  [616]  2.117890884  0.259344621  1.798251538  0.370271455  0.457312303
#>  [621]  0.635490947  1.075874346  0.238009564  4.319575977  0.811118660
#>  [626]  0.656207596  0.778265765  0.610986552  0.629616076  1.080744304
#>  [631]  0.518380038  0.830830397  0.302942345  0.461177349  1.904540317
#>  [636]  0.774632621  4.624218943  1.132922432  0.404283865  1.124651233
#>  [641]  0.240869698  0.817689239  0.998032309  0.935959662  0.423841000
#>  [646]  1.945587111  0.170912357  0.526960441  0.499209407  0.828356768
#>  [651]  0.724232415  1.643340474  1.029107283  3.976591750  0.371199067
#>  [656]  3.632834514  1.260623569  0.763114782  0.833845134  1.566812553
#>  [661]  1.681604435  0.648168299  0.369653048  0.244734745  1.726593576
#>  [666]  1.070231378  0.592047824  0.235999740  0.694703459  1.131376414
#>  [671]  0.304874869  0.331930194  0.401191827  0.149345397  0.599391412
#>  [676]  0.863528691  0.347854186  0.464733192 10.559848166  0.296062562
#>  [681]  0.301087123  0.525646325  0.827197254  0.428710959  0.270939760
#>  [686]  0.705602890  1.181776620  0.574577814  0.320412355  1.055698803
#>  [691]  0.275732418  0.852706561  6.556896798  0.157925800  0.448654599
#>  [696]  0.493798342  0.790633914  0.434276626  0.533917525  0.669116851
#>  [701]  0.338964579  0.139837383  0.336413648  0.208944414  0.632862715
#>  [706]  0.303947257  6.957624820  0.186140640  0.522090482  0.813746892
#>  [711]  0.196344363  0.558808424  0.608822126  0.890661317  0.558731123
#>  [716]  0.289414682  1.629503608  0.495035157  0.472154081  1.128361677
#>  [721]  0.128628748  0.687746376  1.114988617  0.110231126  0.293666234
#>  [726]  0.721604183  0.311213545  0.156843587  0.115023784  0.310440535
#>  [731]  0.796740687  0.461177349  1.019521968  0.620030761  1.303448284
#>  [736]  0.319330142  0.171839968  0.789783603 17.925931101  0.337263958
#>  [741]  0.196421664  3.093969729  1.274460435  0.652110647  0.123294984
#>  [746]  0.392688725  0.226646327  0.086267838  0.596608579  1.757591249
#>  [751]  0.319330142  0.724387016  0.367720525  0.751210439  0.873500511
#>  [756]  0.732117109  0.280911580  1.231403817  0.209872025  0.182198292
#>  [761]  0.072740175  0.394080142  0.336568250  0.197735779  0.139141674
#>  [766]  0.395548860  0.167279213  0.428556357  1.634528168 14.290932161
#>  [771]  0.112627455  0.029838159  0.303096947  1.515020930  2.326989900
#>  [776]  1.204271191  0.161636245  0.089514477  0.201291622  0.664942601
#>  [781]  0.252232935  0.307503100  0.396553772  0.413559976  0.239323680
#>  [786]  2.282464565  0.224249998  0.207552997  0.568161837  0.219380040
#>  [791]  0.278824455  0.742707337  0.213273266  0.245353152  0.126000516
#>  [796]  0.089591778  0.317629522  1.257067726  0.149422698  0.420362458
#>  [801]  0.518689241  0.318866337  0.085804032  0.612377969  0.452751548
#>  [806]  0.098635987  0.131643484  0.105051964  0.219534642  0.059753619
#>  [811]  0.155220268 28.103062361  0.348859098  0.364319284  0.145480351
#>  [816]  1.458127445  0.405366078  0.118579627  0.128319544  0.125768613
#>  [821]  0.346540070  0.490860906  1.896964826  0.049395294  0.932790324
#>  [826]  5.265430158  0.468366336  0.440538001  0.554788776  0.446412872
#>  [831]  0.184362718  0.279597464  0.075600310  0.890042910  0.153751550
#>  [836]  0.316392707  0.187995862  0.272485779  4.065874324  0.128319544
#>  [841]  0.317320318  0.034553516  0.120048345  0.339273782  0.834463541
#>  [846]  1.777998694  1.787197505  0.272485779  0.703593066  0.796044979
#>  [851]  0.171994570  0.219070836  0.169134435  0.197040071  0.685040843
#>  [856] 36.411598237  1.078425277  0.089669079  0.248213287  0.477951651
#>  [861]  0.100491209  0.613150978  2.868714819  0.539328590  0.475168818
#>  [866]  0.167047310  0.501837639  0.482125901  0.811737068  0.054651758
#>  [871]  0.997568504  0.784449839  0.037877456  0.201832729  0.331311787
#>  [876]  0.555252581  0.038341261  6.019268829  0.119971044  0.776642445
#>  [881]  0.959845650  0.384881331  0.332316699  0.625519127  0.191010598
#>  [886]  0.014146070  0.049008790  0.513510079  0.105747672  0.017315408
#>  [891]  0.046303257  0.619644256  0.164419078  0.845131069  0.488696480
#>  [896]  0.217834021  0.083562305  0.410777143  0.231052480  0.237700360
#>  [901]  0.742707337  0.439301186  0.119429937  0.181193380  0.218375128
#>  [906]  0.006261375  3.876950851  0.222781281  0.205543173  3.605238081
#>  [911]  0.033084798  0.182971302  0.220616855  0.097012667  0.117342812
#>  [916]  0.654429675  0.193716131  0.405907184  0.075059203  1.795082200
#>  [921]  0.495421661  0.128551447  0.230434073  0.402351341  1.024623829
#>  [926]  0.386195447  0.316006202  0.085572130  0.763733190  0.803002062
#>  [931]  0.273722594  0.459012923  5.217117077  0.242106513  0.060913133
#>  [936]  1.640016534  0.574809717  0.459708632  0.719130553  0.239787485
#>  [941]  0.310672438  0.353033348  2.092690781  0.122290072  0.093070320
#>  [946]  1.437178893  0.085958634  0.054342554  0.394543948  0.271248964
#>  [951]  0.255015769  0.039732678  0.366561011  0.152669337  0.295521456
#>  [956]  0.969817470  0.190778696  0.190160288  0.178487848  0.959381844
#>  [961]  0.165423991  0.073590486  0.375759821  0.116492502  0.535154339
#>  [966]  0.107912098  2.363166736  0.069648138  0.414874092  0.235845138
#>  [971]  0.028987849  0.085804032  0.070343846  0.275191311 14.744920524
#>  [976]  1.980913636  0.046457859  0.342211218  0.082016287  0.446412872
#>  [981]  1.510150971  6.378872756  0.214819285  0.292661322  0.532216904
#>  [986]  0.122599275  0.367875127  0.440460700  0.521858579  1.138642701
#>  [991]  0.447495085  0.265837899  0.329688467  0.433426315  0.914160800
#>  [996]  0.769839963  0.197735779  2.582933280  0.401810235  0.281220784
#> [1001]  0.782517316  0.324586606  0.276660029  0.453292654  0.483826522
#> [1006]  0.190314890  0.413250773  0.093688727  0.171221560  0.709699840
#> [1011]  0.652497151  0.293202428  0.195416751  0.515906408  0.312063855
#> [1016]  4.740556843  0.842580139  0.146639865  0.606039292  0.143934332
#> [1021]  0.397326781  0.397867888  0.758785930  0.247672180  0.990302216
#> [1026]  0.342133917  0.591429417  0.151509823  0.329688467  3.037308148
#> [1031]  0.204383659  0.097244570  0.362850566  0.187995862  0.573186397
#> [1036]  0.406370990  0.167279213  9.068867825  0.502533347  0.273413390
#> [1041]  0.165810495  0.642216128  0.576201133  0.485217939  0.337341259
#> [1046]  0.267538519  0.509181227  0.119739141  0.244657444  0.703593066
#> [1051]  0.838483189  4.230370704  1.176829361  0.082634694  0.435590741
#> [1056]  0.314228281  0.160244828  0.169675542  0.187918561  0.282148395
#> [1061]  0.726551443  0.770690274  0.306420887  0.310131332  0.188536969
#> [1066]  0.214664683  0.732658216  0.157848499  0.892593840  0.775560232
#> [1071]  0.264832987  0.147180971  0.307193896  0.269934848  0.373363493
#> [1076]  0.282998705  0.525646325  0.407762407  1.537515501  0.205852377
#> [1081]  0.360686140  0.368029728  1.444445181  0.638892188  0.105361168
#> [1086]  0.159935624  1.632904849  0.310285934  0.905580397  0.685813852
#> [1091]  0.771772487  0.550382623  0.753761370  0.406757494  0.380784382
#> [1096]  0.501064629  0.182894001  0.293279729  0.876437946  0.409385726
#> [1101]  1.007153819  0.271017061  0.974455526  0.538246377  0.702510853
#> [1106]  2.459406394  0.183976214  0.338346171  0.128319544  0.876206043
#> [1111]  1.913971031  0.164032574  0.418429935  0.277896844  1.304375895
#> [1116]  0.169366338  0.470298859  0.338732676  0.174390898  0.391761114
#> [1121]  0.178487848  2.520474129  0.336413648  0.521240172  0.161172439
#> [1126]  1.414529721  0.562441568  2.865468180  5.088642931  0.097940279
#> [1131]  0.863374089  2.889586070  0.289801187  5.153498412  2.319646312
#> [1136]  0.529202168  0.199977506  0.558190017  0.260117630  2.917491706
#> [1141]  0.415724402  0.191242501  0.373595395
ei_wgt_sqrt(spec)
#>    [1] 0.94187231 1.46842808 0.95765949 0.72327604 1.01519969 0.70623648
#>    [7] 0.91696225 1.75710279 1.10556746 0.80649528 1.00821526 0.83875937
#>   [13] 0.99865416 0.75817682 0.67555369 1.11647784 1.33480906 0.77600109
#>   [19] 0.65359684 1.22619266 0.79306293 1.40660393 1.05346797 1.38050879
#>   [25] 1.30168857 1.16522252 1.11312822 1.88404957 0.84094827 1.02628710
#>   [31] 0.99340549 0.67529853 0.77080145 0.77784969 1.40353743 1.09811159
#>   [37] 4.79250678 0.82397161 1.47233514 0.93851163 1.22778457 1.09386579
#>   [43] 0.61136037 0.85685174 2.44104450 1.02123672 0.97749038 1.38134082
#>   [49] 3.25613188 0.91765115 2.30949580 1.58949070 0.79646029 0.86173222
#>   [55] 1.00547647 0.88508249 1.13742940 1.05597351 1.11137209 0.74302247
#>   [61] 1.46188000 1.16033128 1.80980178 1.49683027 0.80670895 0.71705385
#>   [67] 0.79746947 0.93446287 0.95308944 0.88819260 1.36548757 0.92364159
#>   [73] 0.74302247 0.50162285 0.76308612 0.80549741 0.90643912 0.85470356
#>   [79] 0.69135629 0.53272580 0.97336767 0.86359687 1.33566956 0.91067558
#>   [85] 1.10286206 0.80955249 0.64732612 0.81040360 0.73258865 1.12022778
#>   [91] 0.72644623 0.61548111 1.52662389 0.64962947 0.96334155 0.91212535
#>   [97] 0.96208841 0.67852355 0.94916339 0.58592975 0.89245131 1.25275085
#>  [103] 0.76601660 0.64171066 0.84496920 0.74849189 0.65911105 0.61697273
#>  [109] 0.86412888 0.94016292 0.75201423 0.61126640 1.10634663 1.33489514
#>  [115] 0.75741872 0.53873095 0.67120267 0.57584210 1.16650369 0.51294759
#>  [121] 1.14808706 0.61529440 0.95188315 0.77266248 0.99784847 0.64153159
#>  [127] 3.03463887 0.69582899 1.02511090 1.13914536 0.62502115 0.62124121
#>  [133] 1.76268479 0.66483878 0.63051192 0.55325120 1.40570512 0.65183655
#>  [139] 1.58493016 1.18099133 0.64572669 0.77895673 1.80722876 0.56139450
#>  [145] 1.67099927 0.73727877 2.98046705 4.73007840 0.59959504 1.17288854
#>  [151] 0.90624897 1.03097846 1.10067208 0.91890238 6.47860166 0.68585005
#>  [157] 0.48980194 4.38010862 2.80277376 0.45198199 0.60036106 1.02011101
#>  [163] 0.41652754 0.35405247 0.66076503 0.56016517 0.71985251 0.61846076
#>  [169] 0.82779742 1.10447569 4.04210979 0.78688152 1.20887595 1.19024617
#>  [175] 0.59603944 0.41195094 1.68816975 1.89016855 1.93554831 0.67461761
#>  [181] 0.44429030 0.73759038 1.99808477 1.62087809 1.08415923 1.30837975
#>  [187] 0.86638631 1.54669986 0.58788742 3.40097258 1.04498117 3.66524804
#>  [193] 1.62565585 4.92676486 2.94094163 1.12549754 1.13995197 1.39442105
#>  [199] 1.20873337 2.33192366 1.66803998 0.76293554 0.82904567 0.76195602
#>  [205] 0.43685777 2.84044030 0.52017648 0.90713602 0.75277777 0.69739590
#>  [211] 0.53894418 0.56719745 0.44377278 0.90390043 0.49447127 0.77095050
#>  [217] 0.97366273 0.62234991 0.66319496 2.18091976 0.64412328 0.50991470
#>  [223] 0.65902388 0.52172036 0.94994990 0.74987217 0.59032538 0.46845938
#>  [229] 0.66811430 0.56343743 1.17274159 1.03886093 0.47261052 2.48035997
#>  [235] 0.24466628 0.78520054 0.92594006 1.30437797 0.38018552 1.64102625
#>  [241] 0.45223613 2.27912268 0.85463635 1.03514922 0.73860222 0.63991767
#>  [247] 0.92438767 0.43156554 0.77466737 0.62087120 0.42876097 0.90027042
#>  [253] 3.46071794 0.80862943 0.60026536 1.46823245 0.91262908 0.67017479
#>  [259] 0.27159562 0.67495815 0.78637029 0.83800561 0.53702204 0.82278549
#>  [265] 0.61641380 1.57215604 0.71368097 0.69243569 4.53590807 0.67987687
#>  [271] 0.33298058 1.18961854 0.83395111 0.81654766 0.63505130 1.41754790
#>  [277] 0.80848733 1.25201690 0.64759231 0.79299049 0.67316838 0.74062175
#>  [283] 0.45628306 0.95501634 1.28882616 0.57324235 0.84053828 0.50470557
#>  [289] 0.57833084 0.77422229 0.53883758 0.57404353 0.64027667 0.56770365
#>  [295] 0.42091794 1.21451795 0.51002735 0.63676778 0.50150831 0.45058168
#>  [301] 1.13206294 0.49481970 0.68004585 0.49109043 0.59786793 0.65649101
#>  [307] 0.38379496 0.79139508 0.51205083 0.80456970 0.60825150 0.51272355
#>  [313] 0.58514485 0.66475237 1.99356561 0.87575150 0.53937039 0.54688037
#>  [319] 0.76203141 0.66974604 0.65447522 0.60919526 0.52719731 0.95729949
#>  [325] 0.55470303 0.56139450 0.29296488 0.55480659 0.58396551 2.04230466
#>  [331] 0.72881483 0.35421469 0.69110696 0.53520035 1.12314711 0.76788923
#>  [337] 0.43593627 0.86903460 0.41721659 0.40363923 0.77399965 0.55800732
#>  [343] 0.69961651 0.64483640 1.06631396 0.88527720 0.80912659 0.57233969
#>  [349] 0.46673935 1.19203068 0.55449586 0.51573992 0.59690626 0.86903460
#>  [355] 1.28274971 0.89668978 1.08240917 0.44904908 0.73634313 0.81033271
#>  [361] 0.28622003 0.43843299 0.52017648 1.14978712 0.55366640 0.63730887
#>  [367] 0.62354877 0.70875368 1.39721977 0.91325835 1.10072428 0.84435707
#>  [373] 1.18356668 0.91155832 0.84231345 1.34633675 2.78203596 2.25769726
#>  [379] 0.65333310 0.54582888 0.69475486 0.82842178 0.92823284 0.99195868
#>  [385] 3.00670063 0.69615916 0.68559871 1.15238235 0.90491676 0.79602739
#>  [391] 1.47864267 1.06042522 0.86883626 3.24368725 1.05847311 1.86815711
#>  [397] 1.56155987 0.81956737 1.00992323 1.18657226 0.80841627 1.01361397
#>  [403] 1.16448275 4.53083911 1.90858841 0.99068374 0.92810905 2.01195257
#>  [409] 0.65823886 0.86173222 0.88456308 1.47856497 1.02410166 0.62225759
#>  [415] 0.86618736 0.92401471 1.70300975 1.05597351 1.38350176 1.54446968
#>  [421] 1.55252032 0.55054478 1.46490281 0.88039673 1.32187779 0.99438811
#>  [427] 1.34385959 1.26888076 0.77696291 0.72335546 0.55408128 0.85045849
#>  [433] 1.24548420 1.02735011 0.76788923 0.89444461 0.55335503 1.14968718
#>  [439] 0.79443808 0.60579085 0.77020497 0.60360576 0.63432718 0.78695452
#>  [445] 0.79559426 1.11899631 0.98118601 0.74294515 0.97866511 1.36540343
#>  [451] 0.62886972 0.70915885 0.61808909 0.90339183 0.83312405 1.83317253
#>  [457] 2.64162518 0.92264589 0.64563772 0.35631688 0.84333588 1.53244563
#>  [463] 0.71585108 0.62318014 0.70907783 1.43786831 0.62059355 0.84005972
#>  [469] 0.78175424 1.52966898 0.66888773 0.85187586 1.25307181 1.15606552
#>  [475] 1.06749858 1.12554859 1.04266963 0.97848899 0.84930936 1.09988889
#>  [481] 0.70769917 0.94758842 0.89051806 0.64839022 0.91136924 0.95717946
#>  [487] 0.97501885 0.60702242 1.08012454 0.83923868 0.84524111 0.70932085
#>  [493] 1.18990825 0.88157051 0.52610648 0.89797022 0.76968267 0.55821319
#>  [499] 0.93261671 0.76773959 0.72517982 0.82187730 0.79811756 0.51729704
#>  [505] 0.88196142 0.74309979 1.27335511 1.32673637 0.76361293 0.67987687
#>  [511] 0.67103146 0.83401999 0.69135629 0.96232723 1.96745931 0.98334999
#>  [517] 0.65210089 0.95957720 1.02857960 0.72049068 1.15382716 0.90485327
#>  [523] 0.98223935 1.00175576 2.33723893 1.61370273 1.70378545 1.52624753
#>  [529] 0.47780885 1.14528147 0.83195097 1.93691315 1.10442367 0.90224645
#>  [535] 0.65157209 0.53915733 1.61170786 1.30477430 1.25645986 1.81338520
#>  [541] 0.54361413 0.54435338 2.02858773 0.95170207 1.18671750 2.21782802
#>  [547] 1.28189850 2.78663708 1.05869019 2.21793163 0.56557456 0.58651774
#>  [553] 0.99091567 0.76669127 3.11315232 1.32113876 1.34403058 1.34761626
#>  [559] 1.36632875 0.87101554 0.72240180 0.47972873 1.67628541 0.94564639
#>  [565] 1.53353240 0.62188819 1.00673267 1.34450067 1.24049263 1.08791495
#>  [571] 0.86598837 0.85517394 0.96810059 3.51441808 0.77295983 0.94296957
#>  [577] 1.18240118 1.48198021 1.76330392 0.95381248 1.18720150 1.45035964
#>  [583] 0.63160435 0.93390940 0.80706494 0.59149204 1.00821526 1.61170786
#>  [589] 0.79905275 1.71815627 1.20220440 1.48585164 1.66724766 1.93242930
#>  [595] 1.39277212 1.35250981 0.82925353 1.45134955 1.10176762 1.47085168
#>  [601] 0.61183003 0.99450365 0.38454266 1.25375932 1.14186541 2.77779951
#>  [607] 0.78644335 0.74108701 1.02213639 1.50544107 1.46062193 1.36131608
#>  [613] 1.05706102 0.78877744 0.81907655 1.77424947 0.62087120 1.63488840
#>  [619] 0.74186180 0.82445952 0.97189104 1.26457231 0.59478512 2.53386488
#>  [625] 1.09800695 0.98760555 1.07554072 0.95296888 0.96738823 1.26743113
#>  [631] 0.87778272 1.11126870 0.67103146 0.82793621 1.68251126 1.07302733
#>  [637] 2.62169453 1.29766616 0.77518632 1.29292051 0.59834818 1.10244526
#>  [643] 1.21796608 1.17948238 0.79371461 1.70054541 0.50402215 0.88501758
#>  [649] 0.86139882 1.10961318 1.03753289 1.56288373 1.23678218 2.43118718
#>  [655] 0.74279048 2.32373019 1.36884919 1.06502015 1.11328304 1.52605932
#>  [661] 1.58097432 0.98153725 0.74124203 0.60312969 1.60198319 1.26125161
#>  [667] 0.93808305 0.59226853 1.01616124 1.29678045 0.67316838 0.70240286
#>  [673] 0.77221624 0.47114959 0.94388297 1.13292531 0.71905401 0.83112192
#>  [679] 3.96179254 0.66336819 0.66897361 0.88391338 1.10883630 0.79826150
#>  [685] 0.63459883 1.02410166 1.32535002 0.92413904 0.69010873 1.25265913
#>  [691] 0.64018694 1.12580376 3.12185025 0.48449515 0.81661801 0.85671763
#>  [697] 1.08405324 0.80342644 0.89084056 0.99727258 0.70980663 0.45590519
#>  [703] 0.70713071 0.55728618 0.96987921 0.67214351 3.21583232 0.52599727
#>  [709] 0.88091860 1.09978442 0.54022181 0.91136924 0.95127943 1.15058627
#>  [715] 0.91130620 0.65587816 1.55629010 0.85778987 0.83773135 1.29505155
#>  [721] 0.43725211 1.01106028 1.28735436 0.40477624 0.66067808 1.03564858
#>  [727] 0.68013032 0.48283226 0.41348210 0.67928513 1.08823175 0.82793621
#>  [733] 1.23100888 0.95999619 1.39190564 0.68894230 0.50538807 1.08347015
#>  [739] 5.16183068 0.70802380 0.54032814 2.14447468 1.37634108 0.98451772
#>  [745] 0.42809051 0.76398901 0.58041314 0.35808601 0.94168931 1.61629948
#>  [751] 0.68894230 1.03764362 0.73930191 1.05668052 1.13944791 1.04316539
#>  [757] 0.64617137 1.35289204 0.55852185 0.52039731 0.32881382 0.76534133
#>  [763] 0.70729317 0.54213260 0.45476968 0.76676620 0.49863629 0.79811756
#>  [769] 1.55868766 4.60885635 0.40915233 0.21059541 0.67120267 1.50062511
#>  [775] 1.85977387 1.33790428 0.49015368 0.36476195 0.54698541 0.99415699
#>  [781] 0.61229934 0.67606374 0.76773959 0.78402903 0.59642485 1.84189519
#>  [787] 0.57733664 0.55542752 0.91896490 0.57103332 0.64376643 1.05068311
#>  [793] 0.56302944 0.60389122 0.43276194 0.36491942 0.68710534 1.36691727
#>  [799] 0.47127150 0.79045082 0.87804447 0.68844180 0.35712211 0.95405337
#>  [805] 0.82033807 0.38289579 0.44234648 0.39515273 0.57123449 0.29801972
#>  [811] 0.48032712 6.46308275 0.72009189 0.73587487 0.46501297 1.47217906
#>  [817] 0.77622316 0.41982465 0.43672625 0.43236351 0.71769451 0.85416568
#>  [823] 1.67916175 0.27096031 1.17748371 2.79756265 0.83436433 0.80919759
#>  [829] 0.90808547 0.81457532 0.52347922 0.64465819 0.33521595 1.15018676
#>  [835] 0.47804925 0.68576628 0.52861202 0.63640680 2.45832823 0.43672625
#>  [841] 0.68677082 0.22662562 0.42241660 0.71013030 1.11369579 1.62565585
#>  [847] 1.62985575 0.63640680 1.02264210 1.08775652 0.50561536 0.57063076
#>  [853] 0.50139374 0.54117805 1.00906961 7.35669522 1.26607059 0.36507681
#>  [859] 0.60740086 0.84285890 0.38647992 0.95465534 2.06493593 0.89534335
#>  [865] 0.84040158 0.49829053 0.86366339 0.84653150 1.09842544 0.28501319
#>  [871] 1.21768304 1.07980537 0.23727572 0.54772011 0.70174825 0.90846497
#>  [877] 0.23872400 2.99112638 0.42228058 1.07441844 1.19443795 0.75635610
#>  [883] 0.70281169 0.96423566 0.53283363 0.14500427 0.26989813 0.87364979
#>  [889] 0.39645902 0.16042766 0.26234251 0.95969693 0.49435508 1.12079175
#>  [895] 0.85228040 0.56901766 0.35242613 0.78138672 0.58602779 0.59439865
#>  [901] 1.05068311 0.80806088 0.42132720 0.51896021 0.56972395 0.09647122
#>  [907] 2.40053499 0.57544291 0.55273176 2.31488738 0.22175690 0.52150008
#>  [913] 0.57264073 0.37973193 0.41762947 0.98626673 0.53659397 0.77674106
#>  [919] 0.33401415 1.63344706 0.85812467 0.43712070 0.58524302 0.77333136
#>  [925] 1.23408513 0.75764623 0.68534728 0.35663919 1.06545160 1.09249944
#>  [931] 0.63784950 0.82599107 2.78469852 0.59988241 0.30089735 1.56130232
#>  [937] 0.92432552 0.82661679 1.03387198 0.59700250 0.67953880 0.72438718
#>  [943] 1.76366227 0.42634237 0.37193624 1.46156558 0.35744370 0.28420579
#>  [949] 0.76579158 0.63496083 0.61566776 0.24301709 0.73813539 0.47636386
#>  [955] 0.66276170 1.20062642 0.53251008 0.53164631 0.51507114 1.19414933
#>  [961] 0.49586350 0.33073010 0.74733972 0.41611357 0.89187177 0.40049580
#>  [967] 1.87417472 0.32174933 0.78527370 0.59207450 0.20757301 0.35712211
#>  [973] 0.32335230 0.63955847 4.68149012 1.71591458 0.26278011 0.71319783
#>  [979] 0.34915072 0.81457532 1.49821133 3.07917848 0.56506645 0.65954671
#>  [985] 0.88942068 0.42688102 0.73945731 0.80912659 0.88072294 1.30093808
#>  [991] 0.81556209 0.62859561 0.70002697 0.80263950 1.16566616 1.06970277
#>  [997] 0.54213260 1.95938368 0.77281117 0.64652690 1.07847448 0.69458946
#> [1003] 0.64126288 0.82082813 0.84802319 0.53186239 0.78373588 0.37316986
#> [1009] 0.50447787 1.02707048 0.98480944 0.66015615 0.53894418 0.87568589
#> [1015] 0.68105883 2.65446848 1.11909899 0.46686242 0.94910286 0.46253552
#> [1021] 0.76848751 0.76901062 1.06199514 0.60673843 1.21324013 0.71311728
#> [1027] 0.93759299 0.47455143 0.70002697 2.12474747 0.55117052 0.38018552
#> [1033] 0.73439007 0.52861202 0.92301940 0.77718470 0.49863629 3.67146531
#> [1039] 0.86426184 0.63748913 0.49644244 0.97702010 0.92544358 0.84924171
#> [1045] 0.70810494 0.63060303 0.86995960 0.42187225 0.60303443 1.02264210
#> [1051] 1.11637492 2.50756446 1.32257296 0.35046456 0.80464110 0.68341661
#> [1057] 0.48803942 0.50219515 0.52850334 0.64759231 1.03919268 1.07029336
#> [1063] 0.67487303 0.67894675 0.52937223 0.56486308 1.04355082 0.48437656
#> [1069] 1.15183385 1.07366961 0.62740638 0.46772300 0.67572375 0.63342087
#> [1075] 0.74495291 0.64856740 0.88391338 0.77851410 1.51172448 0.55314735
#> [1081] 0.73219645 0.73961267 1.46525572 0.97448842 0.39573383 0.48756834
#> [1087] 1.55791347 0.67911596 1.16018274 1.00963877 1.07104456 0.90447226
#> [1093] 1.05847311 0.77755421 0.75231974 0.86299795 0.52138991 0.66024317
#> [1099] 1.14136218 0.78006221 1.22351922 0.63468935 1.20349393 0.89444461
#> [1105] 1.02185532 1.91195665 0.52293021 0.70915885 0.43672625 1.14121117
#> [1111] 1.68667176 0.49377369 0.78863177 0.64269467 1.39240084 0.50173736
#> [1117] 0.83608389 0.70956378 0.50912544 0.76308612 0.51507114 1.93554831
#> [1123] 0.70713071 0.88020095 0.48944994 1.45000310 0.91432711 2.06376712
#> [1129] 2.75019749 0.38154306 1.13282389 2.07243401 0.65631597 2.76766783
#> [1135] 1.85683699 0.88689805 0.54519701 0.91086481 0.62179580 2.08241704
#> [1141] 0.78607802 0.53315698 0.74518422